Fix cargo bays
The lookupCenter of a cargo bay needs to be in the middle of the bay itself (i.e. not 'buried' inside the part), and since this part has two bays, it needs to ModuleCargoBays with different lookupCenters.

How to test cargo bays: put a deployable solar panel inside, and (e.g. with Part Commander) try to deploy the panels while the bay is shut.  If they deploy, the bay isn't shielding them.